pyproject.toml
[project]
name = "mini-e-commerce-order-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.133.0",
    "pydantic[email]>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
